Added BigInteger variable because int is too small for ID

+++ src/main/java/models/pet/CategoryData.java
@@ -3,10 +3,12 @@ package models.pet;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class CategoryData {
-    private int id;
+    private BigInteger id;
     private String name;
 
     @Override
